moving calendar headers to own component
